Stop relying on the lack of gather index validation on GPU
Our tests would run fine on GPU but fail on CPU because
TensorFlow does not validate the indices passed to .gather()
when it runs on the GPU -- invalid indices return 0, a fact
we had been relying on. However, when running on the CPU,
this raises an error. There's no gather_with_default alternative,
or anything of the sort, so we explicitly stop relying on
this weird wrinkle in TensorFlow design.