Fix numerical instability in Shepard Goodness
We were using tf.reduce_sum -- which will produce huge numbers as we're
summing squared differences in rank variables where each rank can be in
the dozens of millions for a moderately-sized dataset -- and then dividing
by n * (n**2 - 1) which can be huge too! Removing one n factor from the
denominator and using tf.reduce_mean instead is equivalent and less
prone to catastrophic division.
ALSO, we were getting bit in the ass by tensorflow's type inference. The
result of tf.argsort is a bunch of int32s and calling reduce_mean() on
them keeps the dtype as integer, even if the result should be
fractional. We now do an explicit cast to avoid this issue.