Add logic to give unique names to metrics in MetricSet
In case a user wants to compute the same metric for different
hparam values (for instance, k) our logic would fail to retain
all values in the output dictionary since the identifier was
only the name of the metric itself. We incorporate the hparams
that are necessary to differentiate the metrics to the dict key
used for it now (see MetricSet._unique_name_for).