Use tf.Variable to make sure tf.function behaves
We were hitting a problem with tf.function over-optimizing our Procrustes
implementation due to a change in the k parameter from 1 to 7 not being
picked up by the graph builder. There was global state (in this case,
instance variables) that our tf.function depended on but couldn't keep
proper track of. The correct way to do it is to use tf.Variables. I expect
I missed some stuff during the migration, but the code works apparently.
If errors pop up, I'll fix them =).

Also updating uv.lock as apparently I forgot to?