Restore tf.constant() use to public functional API
Dropping tf.constant() caused constant retracing whenever users
would call the functional metric API with varying values for the
metric parameters, such as k. This was not the case when we used
tf.constant.
With a change to __setattr__'s inner working in all metrics we
can now uniformly make sure that tf.Variables are being used (and
correctly used) inside the Metric subclasses, and keep the use
of tf.constant in the functional API. Yay!