Split the SVD out of Procrustes to avoid OOM
When dealing with data that is really high-dimensional (say 5000-d),
the SVD procedure creates huge intermediate matrices. We then resort
to tf.map_fn to make sure we process only a few elements at a time.
The conditional is quite empirical, so maybe we'll still hit OOMs,
which would prompt us to change the code.