Add _fn to Metric to avoid repeating measurement code logic
This also allows the .measure() functions to call the tf.functions directly,
without going through the public functional API. As it turns out, in Graph mode,
tf.constant() cannot handle a tf.Variable argument. We just didn't have tests
for that, so it didn't get caught. With this commit, we:
* Keep the functional API clean and working;
* Keep the same public API for the class-based API;
* Avoid repeatedly coding the calls to the _impl functions and
  the subsequent aggregation or not (in the case of LocalizableMetrics)
  of the results returned by the impl.
It took a little time to get right, but now it's nice =).