Extract measure_from_dict into Metric
This means that the implementation is guaranteed to call self.measure().
This also means we need a way to reliably define what items from the dict
must be extracted by measure_from_dict. Right now, there are two options:
by DEFAULT, the keys to be extracted are the names of the parameters of
self.measure. Otherwise, a user of a metric can opt to define custom dict
keys for self.measure. measure_from_dict will then extract those and feed
them to self.measure.